+++ tests/testthat/test-unnest-tokens.R
@@ -14,6 +14,14 @@ test_that("tokenizing by character shingles works", {
   expect_equal(nrow(d), 14)
   expect_equal(ncol(d), 1)
   expect_equal(d$char_ngram[1], "tidy")
+
+  d <- tibble(txt = "Emily Dickinson")
+  d <- unnest_tokens(d, char_ngram, txt,
+                     token = "character_shingles",
+                     to_lower = FALSE, n = 5)
+  expect_equal(nrow(d), 10)
+  expect_equal(ncol(d), 1)
+  expect_equal(d$char_ngram[1], "Emily")
 })
 
 test_that("tokenizing by character shingles can include whitespace/punctuation", {