+++ NEWS.md
@@ -1,6 +1,7 @@
 # tidytext (development version)
 
 * `scale_x/y_reordered()` now uses a function `labels` as its main input (#200)
+* Fixed how `to_lower` is passed to underlying tokenization function for character shingles (#208)
 
 # tidytext 0.3.2